Improve download schema UI
